create isMobile hook
